requirements.txt
asgiref==3.8.1
attrs==25.1.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.3.0
